fix(session): compare exp as float to keep fractional expirations valid
Casting exp to int before the boundary check truncated fractional
NumericDate values, so a token expiring at 1700000000.5 was rejected at
1700000000 even though it was still valid. Compare as float instead and
cover both sides of the fractional boundary in the expiration matrix.

+++ lib/SessionManager.php
@@ -442,7 +442,9 @@ class SessionManager
         if (!isset($decoded['exp']) || !is_numeric($decoded['exp'])) {
             throw new \InvalidArgumentException('JWT exp claim is missing or invalid');
         }
-        if ((int) $decoded['exp'] <= time()) {
+        // Compare as float so fractional NumericDate values are not truncated
+        // before the boundary check.
+        if ((float) $decoded['exp'] <= time()) {
             throw new \InvalidArgumentException('JWT has expired');
         }
